1.检查分析部分 if 判断简化处理等
Former-commit-id: 3cee1ac97596f1bb5efb30d2759b1a6dbb721409

+++ DevLibUtils/src/main/java/dev/utils/app/assist/ScreenSensorAssist.java
@@ -227,7 +227,7 @@ public final class ScreenSensorAssist {
                 // 屏幕旋转时
                 float OneEightyOverPi = 57.29577957855f;
                 float angle = (float) Math.atan2(-Y, X) * OneEightyOverPi;
-                orientation = 90 - (int) Math.round(angle);
+                orientation = 90 - Math.round(angle);
                 // normalize to 0 - 359 range
                 while (orientation >= 360) {
                     orientation -= 360;
@@ -264,7 +264,7 @@ public final class ScreenSensorAssist {
                 // 屏幕旋转时
                 float OneEightyOverPi = 57.29577957855f;
                 float angle = (float) Math.atan2(-Y, X) * OneEightyOverPi;
-                orientation = 90 - (int) Math.round(angle);
+                orientation = 90 - Math.round(angle);
                 // normalize to 0 - 359 range
                 while (orientation >= 360) {
                     orientation -= 360;
